% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{Dire bonjour a quelquun}
\usage{
say_hello(personne)
}
\arguments{
\item{personne}{Character, la personne a qui on souhaite dire bonjour}
}
\value{
Character, de type "Hey coucou {personne}
}
\description{
Dire bonjour a quelquun
}
\examples{
say_hello(personne = "Hédia")
say_hello(personne = "Mickaël")
}
